# internal/handler/static/openapi.yaml
openapi: 3.0.3
info:
  title: Simple Host API
  description: >-
    Static website hosting with a light per-site backend. Deploy a site with one
    API call — inline JSON files (LLM-friendly) or a tar.gz/zip archive — and each
    site gets shared JSON state (with atomic ops), append-only collections. Sign-in is a 6-digit email code or Google that
    returns an API key, sent as the
    `X-API-Key` header. Hosted-page sign-in uses the same account but a
    site-scoped session cookie on the site's own address — never an API key.
    Every site lives at its own address, `https://<site>.<handle>.simple-host.app/`
    (its own browser origin); `https://<handle>.simple-host.app/` lists the
    person's public sites. For a brand-new account a site may briefly live at
    `https://<handle>.simple-host.app/<site>/` until its certificate is issued:
    use the `site_url` the API returns, never a composed one.
    Old `<handle>.simple-host.app/<site>/` links redirect to the site's address,
    as do old `sites.simple-host.app/<handle>/<site>/` ones.
    Visitors sign in with Google or an emailed code on the site's own address (a
    sign-in covers that site only); every save from a page needs a signed-in visitor. Agents save with the API key
    (or the connector) through `https://simple-host.app/v1/...`.
  version: 2.0.0
servers:
  - url: https://simple-host.app
...
paths:
...
  /v1/auth:
    post:
      operationId: postAuth
      summary: Request a sign-in code
      description: >-
        Emails a 6-digit code (and a magic link) to the address. No user row is
        created here — the account is created lazily on the first successful
        verify, so this can't be used to enumerate users.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, example: you@example.com }
                nonce_hash:
                  type: string
                  description: >-
                    Browser pages only. base64url SHA-256 of a random nonce the page keeps. With it the
                    email also carries a sign-in link, redeemable only together with that nonce (so only in
                    that browser). Without it the email carries the code alone.
      responses:
        '202':
          description: Code sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  email: { type: string }
                  expires_in_seconds: { type: integer, example: 900 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/auth/verify:
    post:
      operationId: postAuthVerify
      summary: Verify the code and get an API key
      description: >-
        Exchange `{email, code}` (CLI/agent) or a magic-link `{token}` (browser)
        for an API key. The first successful email-code verify creates the
        account. Dashboard Google also redeems through this endpoint
        (`/?token=&cn=`). A link token is redeemed only with the `nonce` whose hash it
        was issued with, so a sign-in link works only in the browser that asked for it.
        Visitor-issued codes are not accepted here. The code expires in 15 minutes and allows 3 attempts.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                code: { type: string, example: '123456' }
                token: { type: string, description: magic-link token (alternative to email+code) }
                nonce: { type: string, description: required with token — the nonce whose hash was sent when the link was requested }
                name: { type: string, maxLength: 60, description: "Optional label for the key this sign-in issues, shown in the Keys list (GET /v1/me/keys). Defaults to \"dashboard sign-in\" for a link or Google, \"agent sign-in\" for a typed code." }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: Invalid or expired code
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /v1/me:
    patch:
      operationId: patchMe
      summary: Edit display name or change your handle
      description: >-
        Changing the handle changes every address (https://<handle>.simple-host.app/ and
        https://<site>.<handle>.simple-host.app/). Before anything is published it changes freely.
        After that it can change once every 30 days: the old handle stays reserved for this account
        and every old address redirects to the new one. A new handle gets its own certificate; until
        it is issued, sites briefly use the fallback https://<handle>.simple-host.app/<site>/. What pages kept
        in visitors' browsers starts empty at the new address, and visitors sign in again.
      security: [{apiKey: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                display_name: {type: string, maxLength: 100, description: Trimmed; empty clears the name. Never used in URLs.}
                handle: {type: string, pattern: '^[a-z0-9-]{1,39}$', description: "The <handle> in your address https://<handle>.simple-host.app/. Must not be reserved or already taken (handles and free <name>.simple-host.app addresses share one namespace)."}
      responses:
        '200': {description: Updated account profile}
        '400': {description: Invalid profile}
        '401': {description: API key required}
        '409': {description: Handle taken}
        '429':
          description: The handle already changed within the last 30 days after publishing
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  next_change_after: { type: string, format: date-time }
    get:
      operationId: getMe
      summary: Current user
      tags: [Auth]
      security: [{ apiKey: [] }]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  username: { type: string }
                  handle: { type: string, description: "The <handle> in your addresses: sites at https://<site>.<handle>.simple-host.app/, your page at https://<handle>.simple-host.app/" }
                  public_page: { type: string, format: uri, description: "The account's public page listing its public sites, e.g. https://<handle>.simple-host.app/. Present with handle." }
                  display_name: { type: string }
                  is_admin: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/me/api-key/rotate:
    post:
      operationId: postMeApiKeyRotate
      summary: Sign out everywhere (rotate your API key)
      description: The "Sign out everywhere" button. Replaces all of the authenticated owner's API keys with one new key and returns it. Every older key stops working immediately, and every connected app (see /v1/me/connections) is disconnected. Keys are stored only as SHA-256 hashes, so this response is the only time the new key is shown.
      tags: [Auth]
      security: [{ apiKey: [] }]
      responses:
        '200':
          description: API key rotated
          content:
            application/json:
              schema:
                type: object
                required: [api_key, message]
                properties:
                  api_key: { type: string }
                  message: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /v1/me/sign-out:
    post:
      operationId: postMeSignOut
      summary: Sign out (end the key this request used)
      description: >-
        Deletes the API key sent with this request, so it stops working everywhere; the account's
        other keys and connected apps are untouched. The dashboard's Sign out calls this before it
        clears the key from the browser. Needs one of the account's own keys (the admin env key or a
        connected app gets 400 not_an_account_key).
      tags: [Auth]
      security: [{ apiKey: [] }]
      responses:
        '204': { description: Signed out; the key no longer works }
        '400': { description: Not one of the account's own keys (code not_an_account_key), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/me/keys:
    get:
      operationId: getMeKeys
      summary: List your API keys
      description: >-
        Every key on the account, newest first. `current` marks the key this request was made with.
        Keys issued before names existed have no `name` or `last4` (shown as "Earlier key").
        `last_used_at` is updated at most every 5 minutes.
      tags: [Auth]
      security: [{ apiKey: [] }]
      responses:
        '200':
          description: The account's keys
          content:
            application/json:
              schema:
                type: object
                required: [keys]
                properties:
                  keys:
                    type: array
                    items: { $ref: '#/components/schemas/APIKey' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: postMeKey
      summary: Create a named API key
      description: >-
        Mints a new key with the given name (for a CI secret or another machine) and returns it once;
        it is stored only as a hash. Needs one of the account's own keys (the admin env key or a
        connected app gets 400 not_an_account_key).
      tags: [Auth]
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 60, example: GitHub Actions }
      responses:
        '201':
          description: Key created. `api_key` is shown only here.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/APIKey' }
                  - type: object
                    required: [api_key]
                    properties:
                      api_key: { type: string, example: shk_0123abcd… }
                      message: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/me/keys/{id}:
    delete:
      operationId: deleteMeKey
      summary: Revoke one API key
      description: >-
        The key stops working at once; the account's other keys and connected apps are untouched.
        Revoking the key this request uses signs this caller out. Needs one of the account's own keys.
      tags: [Auth]
      security: [{ apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Revoked }
        '400': { description: Not one of the account's own keys (code not_an_account_key) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: No key with that id on this account }
...
  /v1/admin/users:
    post:
      operationId: createEventAccounts
      summary: Create participant accounts (admin only)
      description: >-
        Creates accounts from emails or a count and optional prefix (default guest), numbered
        team-01 through team-30 for count 30 and prefix team. There is no fixed ceiling: ask for
        ten thousand and you get ten thousand, if the server has room for them. How many that is
        comes from its disk and its per-site limit, and GET /v1/admin/usage reports what is being used. A
        request beyond that room is refused with 409 and the real number rather than part-filled.
        Existing usernames are skipped with a reason and their keys are never disclosed. Non-admin
        callers receive 404.
      tags: [Admin]
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [emails]
                  additionalProperties: false
                  properties:
                    emails:
                      type: array
                      minItems: 1
                      items: { type: string, format: email }
                - type: object
                  required: [count]
                  additionalProperties: false
                  properties:
                    count: { type: integer, minimum: 1 }
                    prefix: { type: string, default: guest, pattern: '^[a-z0-9-]{1,39}$' }
      responses:
        '201':
          description: Accounts created and existing usernames skipped
          content:
            application/json:
              schema:
                type: object
                required: [created, skipped]
                properties:
                  created:
                    type: array
                    items:
                      type: object
                      properties:
                        username: { type: string }
                        handle: { type: string }
                        api_key: { type: string }
                        display_name: { type: string, nullable: true }
                  skipped:
                    type: array
                    items:
                      type: object
                      properties:
                        username: { type: string }
                        reason: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    get:
      operationId: getAdminUsers
      summary: List all users + their sites (admin)
      description: >-
        Admin-only. Returns every registered user with their sites nested (API keys
        are never included). Powers the /admin dashboard. A signed-in non-admin gets
        404, not 403 — the endpoint does not confirm it exists to anyone who may not
        use it, and /admin renders that as an ordinary not-found page.
      tags: [Admin]
      security: [{ apiKey: [] }]
      responses:
        '200':
          description: All users with nested sites and totals.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_count: { type: integer }
                  site_count: { type: integer }
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        username: { type: string }
                        handle: { type: string }
                        is_admin: { type: boolean }
                        created_at: { type: string, format: date-time }
                        site_count: { type: integer }
                        suspended: { type: boolean, description: The operator has suspended this account }
                        suspended_reason: { type: string }
                        sites:
                          type: array
                          items:
                            type: object
                            properties:
                              id: { type: string, format: uuid }
                              name: { type: string }
                              site_url: { type: string }
                              active_version: { type: integer }
                              custom_domain: { type: string }
                              created_at: { type: string, format: date-time }
                              suspended: { type: boolean }
                              suspended_reason: { type: string }
                              suspended_by: { type: string, enum: [site, account, ''] }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: not an admin (deliberately indistinguishable from a route that does not exist) }
...
  /v1/admin/users/{id}:
    delete:
      operationId: deleteEventAccount
      summary: Delete an account and its files (admin only)
      description: Deletes the account, cascading database records, its by-id directory and handle symlink. Admin accounts cannot be deleted. Non-admin callers receive 404.
      tags: [Admin]
      security: [{ apiKey: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Account deleted }
        '400': { description: Cannot delete an admin account }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    apiKey: { type: apiKey, in: header, name: X-API-Key }
    connector:
      type: oauth2
      description: >-
        Per-person OAuth 2.1 (authorization code; PKCE S256 required for every self-registered
        client). An access token sent as `Authorization: Bearer` acts exactly as that person's
        X-API-Key on every route here. Discovery at /.well-known/oauth-authorization-server.
      flows:
        authorizationCode:
          authorizationUrl: https://simple-host.app/oauth/authorize
          tokenUrl: https://simple-host.app/oauth/token
          refreshUrl: https://simple-host.app/oauth/token
          scopes:
            sites: Publish and manage your sites and their data
...
  schemas:
...
    APIKey:
      type: object
      required: [id, created_at, last_used_at, current]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, description: "Where it came from (dashboard sign-in, agent sign-in, event account, replacement key) or the name typed when it was created. Absent on keys issued before names existed." }
        last4: { type: string, description: Last 4 characters of the key, to recognise it. Absent on older keys. }
        created_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time, nullable: true }
        current: { type: boolean, description: This is the key the request was made with }
    AuthResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        api_key: { type: string, description: "A new key for this sign-in, starting shk_ (keys issued before 2026-09-27 are bare hex and keep working). Earlier keys keep working until revoked (DELETE /v1/me/keys/{id}) or rotated. Stored only as a hash, so it is shown only here." }
        is_admin: { type: boolean }
        created: { type: boolean, description: true if this verify created the account }
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: What went wrong, in plain words. The wording may change; branch on `code`. }
        code:
          type: string
          description: >-
            A stable snake_case name for the refusal, present where one status
            has more than one meaning. site_exists (409: this account already
            has a site of that name), domain_taken (409: the address belongs to
            another site or person), invalid_name (400: not 1-63 lowercase
            letters, digits and hyphens), name_reserved (400: reserved by the
            platform), invalid_domain (400: not a domain that can be connected),
            site_quota_reached (403), append_only (409: public list items cannot
            be edited; the owner may still delete them), custom_domain_required (409: the site needs
            its own address first), not_an_object (409: the saved value is not a
            JSON object), private_visitor_only and private_needs_own_domain
            (403), use_custom_domain and visitor_auth_required (401),
            missing_api_key, wrong_auth_header (sent Authorization instead
            of X-API-Key) and invalid_api_key (unknown, revoked or signed-out
            key) (401), csrf_required (403), recently_deleted (409: the name
            belongs to a site in Recently deleted), site_not_deleted (409:
            restore of a live site), confirm_required (400),
            not_found (404, private collections), export_link_invalid (404:
            the download link expired or is not valid), site_suspended (403:
            the operator has taken the site down; changes and public reads of
            its data are refused until it is restored, `reason` says why),
            account_suspended (403: the operator has suspended this account;
            its keys, connected apps and sign-in are refused, `reason` says
            why). New codes may be added.
          example: domain_taken
  responses:
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: "Missing or invalid API key — send the X-API-Key header, not Authorization Bearer. The body's code is missing_api_key, wrong_auth_header or invalid_api_key.", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    NotFound: { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    TooManyRequests: { description: Rate limited, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...